Added model json edit and granite example

+++ scripts/zipnn_decompress_file.py
@@ -15,14 +15,14 @@ def check_and_install_zipnn():
         import zipnn
 
 
-def decompress_file(input_file, delete=False, force=False):
+def decompress_file(input_file, delete=False, force=False, hf_cache=False,):
     import zipnn
 
     if not input_file.endswith(".znn"):
         raise ValueError("Input file does not have the '.znn' suffix")
 
     if os.path.exists(input_file):
-        if delete:
+        if delete and not hf_cache:
             print(f"Deleting {input_file}...")
             os.remove(input_file)
         else:
@@ -48,6 +48,21 @@ def decompress_file(input_file, delete=False, force=False):
                 outfile.write(d_data)
                 print(f"Decompressed {input_file} to {output_file}")
 
+            if hf_cache:
+                # If the file is in the Hugging Face cache, fix the symlinks
+                print("Reorganizing Hugging Face cache...")
+                try:
+                    snapshot_path = os.path.dirname(input_file)
+                    blob_name = os.path.join(snapshot_path, os.readlink(input_file))
+                    os.rename(output_file, blob_name)
+                    os.symlink(blob_name, output_file)
+                    
+                    print("To delete the original compressed file, run the script with the --delete flag.")
+                    if os.path.exists(input_file) and delete:
+                        os.remove(input_file)
+                except Exception as e:
+                    raise Exception(f"Error reorganizing Hugging Face cache: {e}")
+
     else:
         print(f"Error: The file {input_file} does not exist.")
 
@@ -65,11 +80,18 @@ if __name__ == "__main__":
     parser.add_argument(
         "--force", action="store_true", help="A flag that forces overwriting when decompressing."
     )
+    parser.add_argument(
+        "--hf_cache",
+        action="store_true",
+        help="A flag that indicates if the file is in the Hugging Face cache.",
+    )
     args = parser.parse_args()
     optional_kwargs = {}
     if args.delete:
         optional_kwargs["delete"] = args.delete
     if args.force:
         optional_kwargs["force"] = args.force
+    if args.hf_cache:
+        optional_kwargs["hf_cache"] = args.hf_cache
 
     decompress_file(args.input_file, **optional_kwargs)
